requirements.txt
Brotli @ file:///D:/bld/brotli-split_1687884251065/work
certifi==2023.5.7
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
faiss==1.7.4
gensim @ file:///D:/bld/gensim_1679970117561/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
lxml==4.9.3
numpy @ file:///D:/bld/numpy_1688887170611/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1689538620473/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-docx @ file:///home/conda/feedstock_root/build_artifacts/python-docx_1622121039670/work
python-dotenv==1.0.0
regex==2023.6.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
scipy==1.11.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open @ file:///home/conda/feedstock_root/build_artifacts/smart_open_split_1673202927732/work/dist
tiktoken==0.4.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1688005749963/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
